Added good working changes
